feat(core): refactor services, add seeding, and introduce DTOs
- Refactor role, season, and team stats services to use uint64 IDs
- Enhance season service with pagination and current season retrieval
- Introduce player and season DTOs with improved validation
- Implement database seeding with fake data and CLI flag support
- Update repository/service structures for player–team relationships
- Improve error handling with consistent not found/conflict responses
- Refine logging and code readability across services

+++ api/dto/role_dto.go
@@ -13,7 +13,7 @@ type UpdateRoleRequest struct {
 }
 
 type RoleResponse struct {
-	ID          uint8     `json:"id"`
+	ID          uint64    `json:"id"`
 	Name        string    `json:"name"`
 	Description string    `json:"description"`
 	CreatedAt   time.Time `json:"created_at"`
@@ -21,6 +21,6 @@ type RoleResponse struct {
 }
 
 type RoleShort struct {
-	ID   uint8  `json:"id"`
+	ID   uint64 `json:"id"`
 	Name string `json:"name"`
 }
