+++ api/dto/user_dto.go
@@ -8,7 +8,7 @@ type CreateUserRequest struct {
 	Name     string `json:"name" binding:"required,min=2,max=35"`
 	LastName string `json:"last_name" binding:"required,min=2,max=35"`
 	Username string `json:"username" binding:"required,min=3,max=50,alphanum"`
-	RoleID   uint8  `json:"role_id,omitempty" binding:"omitempty,gte=0,lte=255"`
+	RoleID   uint64 `json:"role_id,omitempty" binding:"omitempty,gte=0,lte=255"`
 }
 
 type UpdateUserRequest struct {
@@ -18,7 +18,7 @@ type UpdateUserRequest struct {
 	Birthdate  *time.Time `json:"birthdate,omitempty"`
 	ImgProfile *string    `json:"img_profile,omitempty"`
 	ImgBanner  *string    `json:"img_banner,omitempty"`
-	RoleID     *uint8     `json:"role_id,omitempty" binding:"omitempty,gte=0,lte=255"`
+	RoleID     *uint64    `json:"role_id,omitempty" binding:"omitempty,gte=0,lte=255"`
 }
 
 type UserResponse struct {
