+++ api/model/role.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Role struct {
-	ID          uint8          `gorm:"primaryKey" json:"id" form:"id"`
+	ID          uint64         `gorm:"primaryKey" json:"id" form:"id"`
 	Name        string         `gorm:"type:varchar(20);not null;uniqueIndex" json:"name" form:"name" binding:"required,max=20"`
 	Description string         `gorm:"type:varchar(100)" json:"description,omitempty" form:"description"`
 	CreatedAt   time.Time      `gorm:"autoCreateTime" json:"created_at,omitempty"`
